// src/lib/features/browse/shared/domain/enums/browse-enums.ts
// Note: BrowseFilterType has been moved to filtering/domain/filtering-enums.ts
// and is re-exported from this module's index.ts for compatibility

export enum BrowseSortMethod {
  ALPHABETICAL = "alphabetical",
  DATE_ADDED = "date",
  DIFFICULTY_LEVEL = "level",
  SEQUENCE_LENGTH = "length",
  AUTHOR = "author",
  POPULARITY = "popularity",
}